fixes in module due to geolocation #6292

+++ CHT/library/api/fermat-cht-api/src/main/java/com/bitdubai/fermat_cht_api/layer/sup_app_module/interfaces/chat_actor_community/interfaces/ChatActorCommunitySubAppModuleManager.java
@@ -120,7 +120,7 @@ public interface ChatActorCommunitySubAppModuleManager extends ModuleManager, Se
     GeoRectangle getGeoRectangleByLocation(String location)
             throws CantCreateGeoRectangleException;
 
-    Address getAddressByCoordinate(float latitude, float longitude)
+    Address getAddressByCoordinate(double latitude, double longitude)
             throws CantCreateAddressException;
 
     GeoRectangle getRandomGeoLocation() throws CantCreateGeoRectangleException;
